requirements.txt
annotated-types==0.7.0
anthropic==0.50.0
antiorm==1.2.1
anyio==4.9.0
app==0.0.1
appnope==0.1.4
argh==0.31.3
argslib==2.1.0
ark==7.7.0
asttokens==3.0.0
attrs==25.3.0
awsebcli==3.23.3
awslambdaric==3.0.2
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.33.0
azure-identity==1.21.0
bandit==1.8.3
beautifulsoup4==4.13.4
black==25.1.0
blessed==1.21.0
blinker==1.9.0
botocore==1.38.9
cement==2.10.14
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
cheroot==10.0.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
cryptography==44.0.2
db==0.1.1
debugpy==1.8.14
decorator==5.2.1
distlib==0.3.9
distro==1.9.0
docformatter==1.7.5
EditorConfig==0.17.0
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
flake8==7.0.0
flake8-bugbear==24.2.6
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.0
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
ibis==3.3.0
identify==2.6.10
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.18.1
ipython_pygments_lexers==1.1.1
isodate==0.7.2
isort==6.0.1
itsdangerous==2.2.0
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
libexec==1.0.1
mangum==0.19.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.7.0
msal==1.32.0
msal-extensions==1.3.1
mypy==1.9.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
nltk==3.9.1
nodeenv==1.9.1
np==1.0.2
openai==1.23.2
opencv-python==4.7.0.72
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pbr==6.1.1
perplexity-client==0.1.1
pexpect==4.9.0
gunicorn==20.1.0
platformdirs==4.3.7
pluggy==1.5.0
pre-commit==3.7.0
prettier==0.0.7
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyDispatcher==2.0.7
pyflakes==3.2.0
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyttsx3==2.98
pytz==2025.2
pyupgrade==3.15.0
PyYAML==6.0.2
pyzmq==26.4.0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
shortcodes==5.4.0
simplejson==3.20.1
six==1.17.0
snapshot-restore-py==1.0.0
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
SpeechRecognition==3.14.2
SQLAlchemy==2.0.40
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
stevedore==5.4.1
syntext==3.1.2
termcolor==2.5.0
threadpoolctl==3.6.0
tokenize_rt==6.1.0
tomli==2.2.1
tool==0.8.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
untokenize==0.1.1
urllib3==1.26.20
virtualenv==20.30.0
wcwidth==0.2.13
web.py==0.40
Werkzeug==3.1.3
WTForms==3.2.1
boto3
paho-mqtt
spacy
opencv-python
web.py
# Core dependencies with compatible versions for Python 3.13
numpy==1.26.4  # Must be installed FIRST
opencv-python-headless==4.8.1.78  # Compatible with NumPy 1.26.x

# Base Flask requirements with Python 3.13 compatibility



gunicorn==20.1.0
Jinja2==3.1.6
itsdangerous==2.2.0
MarkupSafe==3.0.2
SQLAlchemy==2.0.40
blinker==1.9.0

# AI and ML libraries
anthropic==0.50.0
openai==1.23.2
scikit-learn==1.6.1

nltk==3.9.1

# Data handling and processing
pandas==2.2.3
pydantic==2.11.3
pydantic_core==2.33.1

# HTTP and networking

httpx==0.28.1
httpcore==1.0.8
urllib3==1.26.20

# Utility libraries
pillow==10.0.1
PyYAML==6.0.2
python-dotenv==1.1.0
tqdm==4.67.1
matplotlib==3.7.3

# AWS related
boto3
awsebcli==3.23.3
awslambdaric==3.0.2

# Testing and development
pytest==8.1.1
black==25.1.0
flake8==7.0
verbose
